ecom1-master/ecom1-master/requirements.txt
amqp==2.5.2
billiard==3.6.1.0
bootstrap4==0.1.0
celery==4.3.0
certifi==2019.9.11
chardet==3.0.4
cssutils==1.0.2
dj-database-url==0.5.0
Django==2.2.4
django-crispy-forms==1.7.2
django-heroku==0.3.1
django-inlinecss==0.2.0
django-naomi==0.8
django-sendgrid==1.0.1
future==0.18.0
gunicorn==20.0.0
idna==2.8
importlib-metadata==0.23
instamojo-wrapper==1.2.0
kombu==4.6.6
more-itertools==7.2.0
Pillow==6.1.0
psycopg2==2.8.4
pynliner==0.8.0
python-decouple==3.3
pytz==2019.2
redis==3.3.11
requests==2.22.0
soupsieve==1.9.4
sqlparse==0.3.0
stripe==2.40.0
urllib3==1.25.6
vine==1.3.0
whitenoise==4.1.4
zipp==0.6.0
